+++ PDColorPicker/Classes/UIView+Layout.swift
@@ -11,14 +11,39 @@ import UIKit
 @available(iOS 9.0, *)
 extension UIView {
 
-  /// Provides a convenient method for anchoring the view to other
-  /// anchors within its view hierarchy. All parameters have default values of
-  /// `nil` or `0` so only the desired constraints need be provided.
-  ///
-  /// This method should only be called *after* the view has been added to its superview.
-  ///
-  /// This method sets `translatesAutoresizingMaskIntoConstraints` to `false` on
-  /// the view.
+  /**
+   Provides a convenient method for anchoring the receiver to other
+   anchors within its view hierarchy. All parameters have default values of
+   `nil` or `0` so only the desired constraints need be provided.
+
+   This method should only be called *after* the receiver has been added to its superview,
+   otherwise an exception may be thrown.
+
+   This method sets `translatesAutoresizingMaskIntoConstraints` to `false` on
+   the receiver.
+
+   Example:
+
+   ```
+   // in a UIView subclass:
+   let aView = UIView()
+   addSubview(aView)
+   // sets aView to be positioned 10 points from the
+   //   superview's left anchor
+   aView.anchor(left: viewAnchor, leftConstant: 10))
+   ```
+
+   - parameter left: The anchor to pin the left side of the receiver to.
+   - parameter right: The anchor to pin the right side of the receiver to.
+   - parameter top: The anchor to pin the top side of the receiver to.
+   - parameter bottom: The anchor to pin the bottom side of the receiver to.
+   - parameter leftConstant: The number of points of spacing given to the left side (increases to the right).
+   - parameter rightConstant: The number of points of spacing given to the right side (increases to the left).
+   - parameter topConstant: The number of points of spacing given to the top side (increases down).
+   - parameter bottomConstant: The number of points of spacing given to the bottom side (increases up).
+   - parameter height: The fixed height in points.
+   - parameter width: The fixed width in points.
+  */
   func anchor(left: NSLayoutXAxisAnchor? = nil,
               right: NSLayoutXAxisAnchor? = nil,
               top: NSLayoutYAxisAnchor? = nil,
@@ -28,9 +53,7 @@ extension UIView {
               topConstant: CGFloat = 0,
               bottomConstant: CGFloat = 0,
               height: CGFloat = 0,
-              width: CGFloat = 0,
-              heightMult: CGFloat? = nil,
-              widthMult: CGFloat? = nil) {
+              width: CGFloat = 0) {
 
     translatesAutoresizingMaskIntoConstraints = false