+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # PDColorPicker v0.3.0.9000
 
 * Removed unnecessary layout anchor functions.
+* Improved framework documentation.
 
 # [PDColorPicker v0.3.0](https://github.com/pdil/PDColorPicker/releases/tag/0.3.0)
 Released Monday, August 28, 2017.